housing_agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

pandas==2.2.2
scikit-learn==1.5.1
folium==0.17.0

transformers==4.41.2
torch==2.3.1
langchain==0.2.7
langchain-community==0.2.7


